Alteração do método "LerLinha" para abstrato, cada LeiauteSped deverá implementar a sua leitura específica, e criação de 2 novos testes para leitura retroativa com enum IndTipoFrete.
