+++ tests/FiscalBr.Test/Sped/BlocoC/BlocoCRegistro100Test.cs
@@ -100,6 +100,47 @@ namespace FiscalBr.Test.Sped.BlocoC
             efdFiscalFile.Ler("SpedTestC100.1Bv17.txt");
             Assert.Equal("1862", efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.NumDoc);
             Assert.Equal(IndCodMod.Mod1B, efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.CodMod);
+            Assert.Equal(IndTipoFrete.SemCobrancaFrete, efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.IndFrt);
+        }
+
+        [Fact]
+        public void Teste_Leitura_Isolada_C100_indmod_1B_VNova_2011()
+        {
+            var expectedResult =
+            @"|C100|0|1|F000408|1B|00|001|1862||11012020|11012020|1370|1|0|0|1370|0|0|0|0|1090|196,2|0|0|0|17,99|82,84|0|0|";
+
+            File.WriteAllText("SpedTestC100.1Bv17.txt", expectedResult);
+
+            var efdFiscalFile = new FiscalBr.EFDFiscal.ArquivoEFDFiscalV2(
+                null, 
+                new DateTime(2011,12,1),
+                new DateTime(2011,12,31)
+                );
+
+            efdFiscalFile.LerArquivo("SpedTestC100.1Bv17.txt");
+            Assert.Equal("1862", efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.NumDoc);
+            Assert.Equal(IndCodMod.Mod1B, efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.CodMod);
+            Assert.Equal(IndTipoFrete.ContaTerceirosPre2012, efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.IndFrt);
+        }
+
+        [Fact]
+        public void Teste_Leitura_Isolada_C100_indmod_1B_VNova_2015()
+        {
+            var expectedResult =
+            @"|C100|0|1|F000408|1B|00|001|1862||11012020|11012020|1370|1|0|0|1370|0|0|0|0|1090|196,2|0|0|0|17,99|82,84|0|0|";
+
+            File.WriteAllText("SpedTestC100.1Bv17.txt", expectedResult);
+
+            var efdFiscalFile = new FiscalBr.EFDFiscal.ArquivoEFDFiscalV2(
+                null,
+                new DateTime(2015, 12, 1),
+                new DateTime(2015, 12, 31)
+                );
+
+            efdFiscalFile.LerArquivo("SpedTestC100.1Bv17.txt");
+            Assert.Equal("1862", efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.NumDoc);
+            Assert.Equal(IndCodMod.Mod1B, efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.CodMod);
+            Assert.Equal(IndTipoFrete.ContaEmitentePre2018, efdFiscalFile.BlocoC?.RegC001?.RegC100s?.FirstOrDefault()?.IndFrt);
         }
 
         private void SetCulture()
